Product Delete Refactoring

+++ Business/Concrete/ProductImageManager.cs
@@ -13,29 +13,31 @@ namespace Business.Concrete
     public class ProductImageManager : IProductImageService
     {
         private readonly IProductImageDal _productImageDal;
+        private readonly IProductDal _productDal;
         private readonly ICloudinaryService _cloudinaryService;
 
-        public ProductImageManager(ICloudinaryService cloudinaryService, IProductImageDal productImageDal)
+        public ProductImageManager(ICloudinaryService cloudinaryService, IProductImageDal productImageDal, IProductDal productDal)
         {
             _cloudinaryService = cloudinaryService;
             _productImageDal = productImageDal;
+            _productDal = productDal;
         }
 
         [SecuredOperation("Company")]
-        public IResult Upload(IFormFile file,int productId)
+        public IResult Upload(IFormFile file, int productId)
         {
             string filePath = FileHelper.Add(file).Message;
             var uploadResult = _cloudinaryService.Upload(filePath);
             string uploadedImagePath = uploadResult.Message;
             if (!uploadResult.Success) return uploadResult;
-            
+
             var deleteResult = FileHelper.Delete(filePath);
             if (!deleteResult.Success) return deleteResult;
-            
-            ProductImage productImage = new ProductImage { ProductId = Int32.Parse(productId.ToString()),ImagePath = uploadedImagePath,Date = DateTime.Now};
+
+            ProductImage productImage = new ProductImage
+                { ProductId = Int32.Parse(productId.ToString()), ImagePath = uploadedImagePath, Date = DateTime.Now };
             _productImageDal.Add(productImage);
             return new SuccessResult(uploadedImagePath);
-
         }
 
         public IResult Add(ProductImage productImage)
@@ -44,10 +46,17 @@ namespace Business.Concrete
             return new SuccessResult();
         }
 
+        public IResult Delete(ProductImage productImage)
+        {
+            _productImageDal.Delete(productImage);
+            return new SuccessResult();
+        }
+
         [SecuredOperation("Company")]
         public IResult Update(IFormFile file, int imageId)
         {
             ProductImage productImage = _productImageDal.Get(pi => pi.Id == imageId);
+            if (productImage == null) return new ErrorResult(Messages.ImageNotFound);
             var destroyResult = Destroy(productImage.Id);
             if (!destroyResult.Success) return destroyResult;
             var uploadResult = Upload(file, productImage.ProductId);
@@ -61,25 +70,21 @@ namespace Business.Concrete
             ProductImage productImage = _productImageDal.Get(pi => pi.Id == imageId);
             if (productImage == null) return new ErrorResult(Messages.ProductNotHaveImage);
             string publicId = productImage.ImagePath.Split('/')[7].Split('.')[0];
-            var result = _cloudinaryService.Destroy(publicId);
-            if (!result.Success) return result;
+            IResult result = new SuccessResult();
+            if (publicId != "default")
+            {
+                result = _cloudinaryService.Destroy(publicId);
+                if (!result.Success) return result;
+            }
             _productImageDal.Delete(productImage);
             return result;
         }
 
         public IDataResult<ProductImage> GetByProductId(int productId)
         {
+            
             var result = _productImageDal.Get(pi => pi.ProductId == productId);
-            if (result == null)
-            {
-                result = new ProductImage
-                {
-                    Id = -1,
-                    Date = DateTime.Now,
-                    ImagePath = "https://res.cloudinary.com/dbvephcae/image/upload/v1665952284/default.png",
-                    ProductId = productId
-                };
-            }
+            if (result == null) return new SuccessDataResult<ProductImage>(Messages.ProductNotHaveImage);
             return new SuccessDataResult<ProductImage>(result);
         }
 
@@ -88,7 +93,7 @@ namespace Business.Concrete
         {
             return new SuccessDataResult<ProductImage>(_productImageDal.Get(pi => pi.Id == id));
         }
-        
+
         private IResult CheckIfProductHaveImage(int productId)
         {
             var result = _productImageDal.Get(pi => pi.ProductId == productId);
